requirements.txt
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.8.1.78
pandas==2.0.1
python-dotenv==1.0.0
streamlit==1.29.0
tensorflow==2.14.0
visualkeras==0.0.2
Pillow==10.1.0
